// bridge/typescript/src/_generated/dir/symbol/key.ts
export type StaticKey =
    /** Regular name key (like `x` or `"weird identifier"`). */
    | {
          readonly kind: "name";
          readonly name: StringId;
      }
    /** Positional index key (like `0` or `1`). */
    | {
          readonly kind: "index";
          readonly index: number;
      }
    /** Symbol key. */
    | {
          readonly kind: "symbol";
          readonly symbol: SymbolKey;
      }
;

export const StaticKey = {
    ...StaticKeyImpl,

    /** Regular name key (like `x` or `"weird identifier"`). */
    name(name: StringId): StaticKey {
        return { kind: "name", name };
    },

    /** Positional index key (like `0` or `1`). */
    index(index: number): StaticKey {
        return { kind: "index", index };
    },

    /** Symbol key. */
    "symbol"(symbol_: SymbolKey): StaticKey {
        return { kind: "symbol", symbol: symbol_ };
    },

    /** Encode this value. */
    encode(writer: BinaryWriter, value: StaticKey): void {
        encodeStaticKey(writer, value);
    },

    /** Decode one StaticKey. */
    decode(reader: BinaryReader): StaticKey {
        return decodeStaticKey(reader);
    },

    /** Return this value as JSON. */
    toJson(value: StaticKey): Json {
        return toJsonStaticKey(value);
    },

    /** Return one StaticKey from one JSON value. */
    fromJson(value: Json): StaticKey {
        return fromJsonStaticKey(value);
    },
};
// ...
export type SymbolKey =
    /** Unique symbol key from a declaration. */
    | {
          readonly kind: "unique";
          readonly unique: GlobalSymbolId;
      }
    /** Symbol.for registry key (string is the content of `Symbol.for`). */
    | {
          readonly kind: "registry";
          readonly registry: StringId;
      }
;

export const SymbolKey = {
    /** Unique symbol key from a declaration. */
    "unique"(unique_: GlobalSymbolId): SymbolKey {
        return { kind: "unique", unique: unique_ };
    },

    /** Symbol.for registry key (string is the content of `Symbol.for`). */
    registry(registry: StringId): SymbolKey {
        return { kind: "registry", registry };
    },

    /** Encode this value. */
    encode(writer: BinaryWriter, value: SymbolKey): void {
        encodeSymbolKey(writer, value);
    },

    /** Decode one SymbolKey. */
    decode(reader: BinaryReader): SymbolKey {
        return decodeSymbolKey(reader);
    },

    /** Return this value as JSON. */
    toJson(value: SymbolKey): Json {
        return toJsonSymbolKey(value);
    },

    /** Return one SymbolKey from one JSON value. */
    fromJson(value: Json): SymbolKey {
        return fromJsonSymbolKey(value);
    },
};