// bridge/typescript/src/_generated/dir/symbol/symbol.ts
export type SymbolKind = "variable" | "import" | "class" | "struct" | "interface" | "newtypeInterface" | "enum" | "enumField" | "function" | "label" | "extension" | "typeAlias" | "genericTypeParameter" | "genericValueParameter" | "associatedType" | "associatedConst" | "newtype";

export const SymbolKind = {
    /** Encode this value. */
    encode(writer: BinaryWriter, value: SymbolKind): void {
        encodeSymbolKind(writer, value);
    },

    /** Decode one SymbolKind. */
    decode(reader: BinaryReader): SymbolKind {
        return decodeSymbolKind(reader);
    },

    /** Return this value as JSON. */
    toJson(value: SymbolKind): Json {
        return toJsonSymbolKind(value);
    },

    /** Return one SymbolKind from one JSON value. */
    fromJson(value: Json): SymbolKind {
        return fromJsonSymbolKind(value);
    },
};